backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
starlette==0.27.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Auth
authlib==1.2.1
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
itsdangerous==2.1.2

# HTTP
httpx==0.25.2

# Monitoring
prometheus-client==0.22.1
psutil==5.9.6

# Email
mailjet-rest==1.3.4